fix: window resize behaviour

+++ src/view/show_mode/editor/nodes/impl/effects.py
@@ -136,7 +136,7 @@ class AutoTrackerNode(FilterNode):
             trackers = f.number_of_concurrent_trackers + 1
             if trackers < len(self.terminals) / 3:
                 self.terminals.clear()
-            for i in range(len(self.terminals) / 3, trackers, 1):
+            for i in range(int(len(self.terminals) / 3), trackers, 1):
                 min_brightness_filter_id: str = f.get_min_brightness_filter_id(i)
                 self.addOutput(min_brightness_filter_id)
                 self.addOutput("Tracker{}_Pan".format(i))
